Split tests for TrustedHostsNetworkResolver
